[Lib][ML] Move template specialization of constructor of DataLoader to data_loader.cpp Resolved issue #126
